[refactor] domain 내부에 있는 레포지토리(쿠폰,상품,재고,발행된쿠폰,유저)를 domain 외부의 infrastructure로 변경 , ServerApplication에 jpaRepository 설정 추가
