+++ src/main/java/kr/hhplus/be/server/ServerApplication.java
@@ -2,8 +2,10 @@ package kr.hhplus.be.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
+@EnableJpaRepositories("kr.hhplus.be.server.infrastructure")
 public class ServerApplication {
 
 	public static void main(String[] args) {
