+++ src/main/java/kr/hhplus/be/server/infrastructure/coupon/CouponJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.coupon.infrastructure;
+package kr.hhplus.be.server.infrastructure.coupon;
 
 import kr.hhplus.be.server.domain.coupon.Coupon;
 import org.springframework.data.jpa.repository.JpaRepository;
