+++ src/main/java/kr/hhplus/be/server/infrastructure/product/ProductJpaRepository.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.product.infrastructure;
+package kr.hhplus.be.server.infrastructure.product;
 
 import kr.hhplus.be.server.domain.product.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
