+++ src/main/java/kr/hhplus/be/server/infrastructure/user/UserJpaRepository.java
@@ -1,8 +1,7 @@
-package kr.hhplus.be.server.domain.user.infrasturcture;
+package kr.hhplus.be.server.infrastructure.user;
 
 import kr.hhplus.be.server.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-
 
 public interface UserJpaRepository extends JpaRepository<User, Long> {
