+++ src/main/java/kr/hhplus/be/server/infrastructure/user/UserRepositoryImpl.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.user.infrasturcture;
+package kr.hhplus.be.server.infrastructure.user;
 
 import kr.hhplus.be.server.domain.user.User;
 import kr.hhplus.be.server.domain.user.UserRepository;
@@ -8,11 +8,12 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 import java.util.Optional;
 
-@RequiredArgsConstructor
+
 @Repository
+@RequiredArgsConstructor
 public class UserRepositoryImpl implements UserRepository {
 
-    private UserJpaRepository userJpaRepository;
+    private final UserJpaRepository userJpaRepository;
 
     @Override
     public Optional<User> findById(Long id) {
